# AlamiaAccounts-Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AlamiaAccounts API",
        "description": "API testing",
        "version": "1.0.0"
    },
    "paths": {
...
        "/vouchers": {
            "get": {
                "tags": [
                    "Vouchers"
                ],
                "summary": "Get all vouchers",
                "operationId": "5420c51b3fde8ec6228411a29cdbfc5d",
                "parameters": [
                    {
                        "name": "from_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Vouchers"
                ],
                "summary": "Create a new journal voucher",
                "operationId": "2df851426b55bb32b1f285c76fa3476d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date",
                                    "reference",
                                    "currency",
                                    "entries"
                                ],
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-10-01"
                                    },
                                    "reference": {
                                        "type": "string",
                                        "example": "JV-001"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Monthly adjustment"
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "entries": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "account_code": {
                                                    "type": "string",
                                                    "example": "1000"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "example": 100.5
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "enum": [
                                                        "debit",
                                                        "credit"
                                                    ]
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Entry description"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Voucher created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/vouchers/{reference}": {
            "get": {
                "tags": [
                    "Vouchers"
                ],
                "summary": "Get voucher details",
                "operationId": "a7aa7fb326674a06236d0b9dec1634c2",
                "parameters": [
                    {
                        "name": "reference",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Voucher not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Vouchers"
                ],
                "summary": "Delete a voucher",
                "operationId": "e46cb23f2a0a000aacab00f4c1d529f5",
                "parameters": [
                    {
                        "name": "reference",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Voucher deleted successfully"
                    },
                    "404": {
                        "description": "Voucher not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
}